+++ CHANGELOG.md
@@ -5,9 +5,11 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 ### Added
+- Added PyTorch CyclicLR scheduler
 ### Changed
 ### Deprecated
 ### Removed
+- Cyclic LR callback in favour of torch cyclic lr scheduler
 ### Fixed
 - Fixed bug where aggregate predictions couldn't handle empty list
 - Fixed a bug where Runtime Errors on forward weren't handled properly